Gusturoi-1-Initial configurations
