+++ src/main/java/com/sef/App.java
@@ -1,38 +1,11 @@
 package com.sef;
 
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
+import com.sef.gui.GUI;
 
-import java.io.IOException;
-
-public class App extends Application
+public class App
 {
-    private static Scene scene;
-
-    @Override
-    public void start(Stage stage) throws IOException
-    {
-        scene = new Scene(loadFXML("primary"), 640, 480);
-        stage.setScene(scene);
-        stage.show();
-    }
-
-    static void setRoot(String fxml) throws IOException
-    {
-        scene.setRoot(loadFXML(fxml));
-    }
-
-    private static Parent loadFXML(String fxml) throws IOException
-    {
-        FXMLLoader fxmlLoader = new FXMLLoader(App.class.getResource(fxml + ".fxml"));
-        return fxmlLoader.load();
-    }
-
     public static void main(String[] args)
     {
-        launch();
+        GUI.main(args);
     }
 }
