+++ src/main/java/com/sef/gui/GUI.java
@@ -1,0 +1,38 @@
+package com.sef.gui;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+
+public class GUI extends Application
+{
+    private static Scene scene;
+
+    @Override
+    public void start(Stage stage) throws IOException
+    {
+        scene = new Scene(loadFXML("primary"), 640, 480);
+        stage.setScene(scene);
+        stage.show();
+    }
+
+    public static void setRoot(String fxml) throws IOException
+    {
+        scene.setRoot(loadFXML(fxml));
+    }
+
+    private static Parent loadFXML(String fxml) throws IOException
+    {
+        FXMLLoader fxmlLoader = new FXMLLoader(GUI.class.getResource("/com/sef/" + fxml + ".fxml"));
+        return fxmlLoader.load();
+    }
+
+    public static void main(String[] args)
+    {
+        launch();
+    }
+}
