+++ src/main/java/com/sef/gui/controllers/PrimaryController.java
@@ -1,6 +1,9 @@
-package com.sef;
+package com.sef.gui.controllers;
 
 import java.io.IOException;
+
+import com.sef.gui.GUI;
+
 import javafx.fxml.FXML;
 
 public class PrimaryController
@@ -8,6 +11,6 @@ public class PrimaryController
     @FXML
     private void switchToSecondary() throws IOException
     {
-        App.setRoot("secondary");
+        GUI.setRoot("secondary");
     }
 }
