+++ src/main/java/com/sef/gui/controllers/SecondaryController.java
@@ -1,6 +1,9 @@
-package com.sef;
+package com.sef.gui.controllers;
 
 import java.io.IOException;
+
+import com.sef.gui.GUI;
+
 import javafx.fxml.FXML;
 
 public class SecondaryController
@@ -8,6 +11,6 @@ public class SecondaryController
     @FXML
     private void switchToPrimary() throws IOException
     {
-        App.setRoot("primary");
+        GUI.setRoot("primary");
     }
 }
