+++ src/main/java/module-info.java
@@ -4,6 +4,7 @@ module com.sef
     requires javafx.fxml;
     requires com.fasterxml.jackson.databind;
 
-    opens com.sef to javafx.fxml;
+    opens com.sef.gui to javafx.graphics;
+    opens com.sef.gui.controllers to javafx.fxml;
     exports com.sef;
 }
