+++ src/main/resources/com/sef/secondary.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.geometry.Insets?>
 
-<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sef.SecondaryController">
+<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sef.gui.controllers.SecondaryController">
     <children>
         <Label text="Secondary View" />
         <Button fx:id="secondaryButton" text="Switch to Primary View" onAction="#switchToPrimary" />
